Create package ebcmeasurements and move Signed-off-by: Jun Jiang <jun.jiang@eonerc.rwth-aachen.de>
