+++ ebcmeasurements/Base/Auxiliary.py
