+++ ebcmeasurements/Beckhoff/_config/AdsReadDataExamples.json
