+++ ebcmeasurements/Beckhoff/_config/AdsWriteDataExamples.json
